main/requirements.txt
ultralytics
opencv-python-headless
flask
python-dotenv
paho-mqtt
numpy